Reclassify flows upon session modification (#15)
This adds generation id to the session which is checked against the
flow. If the generation id does not match, the flow is re-classified.

In order to enable flow re-classification in case of app detection,
the URI that was extracted by ADF code is now stored in the flow.

Formely, the traffic on flows for modified sessions could result
in crashes such as this:

/src/vpp/src/plugins/upf/upf_input.c:152 (upf_input) assertion `(((upf_buffer_opaque_t *)((u8 *)((b)->opaque2) + ((uword) & (((vnet_buffer_opaque_t *) 0)->unused))))->gtpu.pdr_idx) < vec_len (active->pdr)' fails

Fixes #14.