Fix flow expiration and rescheduling (take 2)
The previous fix (#67) has a bug in it causing multiple rechecks
for flow expirations around the flowtable timer wheel.
Some redundant reshuffling of the per-wheel-entry flow lists has been
removed.
Besides, there was a problem in the flow node that was causing the
flow expirations to be incorrectly rescheduled: flow_update_lifetime()
was being invoked before flow_update_active() for the same flow.