Fix forwarding policy handling for the proxy output
This directs all of the TCP connection packets for both active and
passive proxy connections through the UPG subgraph. Care is taken not
do traffic accounting twice on the same packets. This also applies to
RSTs that are sent by the proxy to terminate a particular connection
upon an error. TCP frames not related to a particular connection, such
as RSTs sent in response to late frames after the connection is
already gone, still don't go through the subgraph at the moment, but
fixing that part would be more problematic as it would require flow
lookup without having SEID.

Newer VPP versions (21.10) does not have separate half-open and normal
connection pools and uses `tcp[46]-output` nodes for all of the
connection-related packets, which is followed by the node that can be
specified per-connection, so there will be a possibility to simplify
the approach. At the moment, only the part that adds next node index /
opaque value is backported (travelping/vpp-base#2), and the opaque
value storing the flow id is obtained by looking up the TCP
connection.