Remove related flows together with PFCP sessions
This is needed to avoid losing IPFIX reports for the flows that still
exist when a PFCP session is deleted.
As part of the change, flowtable hooks are replaced with flowtable
events which are more flexible (multiple handlers per event).
Possible inconsistencies related to flows that don't correspond to any
active session are now being avoided.