rework the upf timer wheel
Decouple memory management for the timer handlers from the start/stop
logic.

Handles must now be allocated before they can be started and they are not
longer automatically freed when stoped. Reusing handles for restarting
or changing timer is now possible.

This solves the problem that handles could accidentally reused.