style: prefix types and functions related to pfcp for consistency (#373)
Done to clarify usage, to avoid confusion and to not poison global namespace with conflicting names.

Not all kinds of names renamed, since PR already quite large and it would be easy to miss error.
No changes in logic, only renames. Except typedef u32 pfcp_timer_t; was removed, since it was a leftover and proper type was pfcp_timer_ie_t which is now named pfcp_ie_timer_t for consistency.