remove pfcp-server-mode since it didn't worked
It impossible to change interrupt mode of process node.

Instead for tests use `use-private-rx-mqs` parameter for session node
which should enable interrupt mode for session processing. Which is
only pooling node we using. In current upg version this is broken due
to bug which is fixed in vpp commit 5f30518 and included in v23.06.

For flowtable timer wheel expiration we using separate process which
sends interrupts to worker input nodes to trigger expiration.