Disable "address-of-packed-member" compiler warnings
This warning spammed a lot and mostly related to ip address alignment in
packed structures which is part of VPP and we can't control it much.
For now disable it, because spam of this warning makes it impossible to
analyze other warnings which has a lot of really important ones.